dpaa2-mac: return -EPROBE_DEFER from dpaa2_mac_open in case the fwnode is not set
commit d726511a349319ff6bde3b0a722b21ddbeb7708e from
https://source.codeaurora.org/external/imx/linux-imx

We could get into a situation when the fwnode of the parent device is
not yet set because its probe didn't yet finish. When this happens, any
caller of the dpaa2_mac_open() will not have the fwnode available, thus
cause problems at the PHY connect time.

Avoid this by just returning -EPROBE_DEFER from the dpaa2_mac_open when
this happens.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>