dpaa2-mac: take into account a defer probe request
commit 5158161e556d52e5487696ca559be419e2aa8948 from
https://source.codeaurora.org/external/imx/linux-imx

Take into account the new defer request which is returned by the
dpaa2_mac_open() and actually defer probe if this happens.

Also, while at it, do not call _open and _get_attributes on the DPMAC
object since it's not necessary. This is already done by the
dpaa2_mac_open() function which is called.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>