octeontx2-bphy-netdev: fix possibility of processing the mbt multiple times
commit e296e6c952d73c485868fe04c8856e77f3273628 from
git@git.assembla.com:cavium/WindRiver.linux.git

The mbt ring is the same for each rx packet type across LMACs in a RFOE,
and the mbt sw pointers are maintained per LMAC, which causes problem when
the napi poll routine is scheduled for different LMACs in the same RFOE at
different circumstances. This patch maintains mbt sw pointers in rfoe common
structure which is shared by all LMACs in a RFOE mhab.

Change-Id: I43c7182c02d40c9614845dcf2e9e036221784547
Signed-off-by: Baha Mesleh <baha.mesleh@nokia.com>
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/65974
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>