dpaa2-mac: only issue phylink_stop/start if necessary
commit 5158161e556d52e5487696ca559be419e2aa8948 from
https://source.codeaurora.org/external/imx/linux-imx

With the dpaa2_mac standalone driver there is a situation that we get
into in which the LINK_UP_REQ irq is triggered multiple times even
though not necessary.
This leads to multiple phylink_start(), for example, calls being made
and a warning splat to be shown.
Just keep an internal field so that we know when the calls are necessary
and when not.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>