spi: spi-cadence-quadspi: Add support for OSPI device reset
commit d3ef2edddad1694516f634e563321677c7d118d3 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

This patch add OSPI flash device reset.
PMC_GPIO12 is used for OSPI device reset. As part of reset sequence
configure the pin to disable the tristate, enables hysteresis and set
the direction of the pin to output before toggling the pin.
Delay of 1Micro second is added in between high and low to meet the
setup and hold time requirement of the flash device.

Signed-off-by: Sai Krishna Potthuri <lakshmi.sai.krishna.potthuri@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>