LF-5113-2 arm64: dts: imx8mp-evk: configure multiple queues on eqos
commit 63e900df71cfa778a3499862efb0b2122fe5dadf from
https://source.codeaurora.org/external/imx/linux-imx

Eqos ethernet support five queues on hardware, enable these queues and
configure the priority of each queue. Uses Strict Priority as scheduling
algorithms to ensure that the TSN function works.

The priority of each queue is a bitmask value that maps VLAN tag priority
to the queue. Since the hardware only supports five queues, this patch maps
priority 0-4 to queues one by one, and priority 5-7 to queue 4.

The total fifo size of 5 queues is 8192 bytes, if enable 5 queues with
store-and-forward mode, it's not enough for large packets, which would
trigger fifo overflow frequently. This patch set DMA to thresh mode to
enable all 5 queues.

Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>
Signed-off-by: Joakim Zhang <qiangqing.zhang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>