LF-5148 drm/mxsfb: Force full mode set if primary plane's framebuffer pitch is changed
commit 8b1e5e8f582e9dcca6261d8296f90acd902fb2b1 from
https://github.com/nxp-imx/linux-imx.git

We use the pigeon registers to do horizontal crop on framebuffer as the
below patch supports.  When primary plane's framebuffer pitch is changed,
we have to do a full modeset as the relevant LCDIF registers, like the
pigeon registers, are not shadowed.  This makes the transition from
bootlogo/framebuffer console stage(where crop is usually not needed) to
GUI stage(where horizontal crop is likely needed) safe.

This patch fixes the below patch:
LF-2321-10: drm/mxsfb: Add support for horizontal stride

Cc: Sandor Yu <Sandor.yu@nxp.com>
Cc: Wujian sun <wujian.sun_1@nxp.com>
Cc: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>