LF-5167 drm/bridge: nwl-dsi: Workaround HX8394F panel display shift issue
commit 35dc5c7657ffc28fea908c2b39148cc8c77a48b8 from
https://github.com/nxp-imx/linux-imx

Similar to the RM68200 panel, running an on-screen multimedia test case
for a while, there will be horizontal display content shift observed at
the HX8394F display panel.  Display color is likely wrong at the same
time.  By tuning the HFP/HBP/HSA settings, the issue is not observed
any more.  This is based on previous knowledge got from support task
for custom external DSI bridges - change the values to be in bytes and
subtract certain magic values.  We have formulas to determine the settings
and other DSI host/DPHY settings from the design team, but we need to
make sure it works well across all impacted platforms.  So, this patch
implements the workaround for now.

Reported-by: Jared Hu <jared.hu@nxp.com>
Cc: Sandor Yu <Sandor.yu@nxp.com>
Cc: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>