arm64: dts: ti: k3-j721s2-common-proc: Add support for CAN instances 3 and 5 in main domain
commit 6f12e9201afb50f472d48383f2d9be20e8aee697 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

CAN instances 3 and 5 in main domain are brought on the Common processor
board through header J27 and J28. The CAN High and Low lines from the SoC
are routed through a mux on the SoM. The select lines need to be set for
the CAN signals to get connected to the transceivers on the common
processor board. Threfore, add respective mux, transceiver dt nodes to add
support for these CAN instances.

Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>