clk: Use clamp instead of open-coding our own
commit  7b9c04e5f346b48675e53754a1662aed295ad538 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

The code in clk_set_rate_range() will, if the current rate is outside of
the new range, force it to the minimum or maximum.

Since it's running under the condition that the rate is either lower
than the minimum, or higher than the maximum, this is equivalent to
using clamp, while being less readable. Let's switch to using clamp
instead.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>