LF-5320: drm/bridge: nxp-seiko-43wvfig: fix boot freeze
commit 80409490d038dd395225e3fdca9f745ef1dbffc7 from
https://github.com/nxp-imx/linux-imx

The following:

added a check in mxsfb driver that retrieved the bus_format from the
bridge itself. However, the nxp-seiko-43wvfig driver did not provide a
hook for retrieving the input formats and the boot process hung with a
NULL pointer exception since there was no bridge state to retrieve.

The fix is to add the atomic_get_input_bus_fmts() hook, together with
the required duplicate_state(), destroy_state() and reset(), which are
necessary when any atomic hook is implemented.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Tested-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>