ASoC: xilinx: xlnx_formatter_pcm: Handle sysclk setting
commit 704bc1ef203d815f9e5d6580498f9ced5306d3de from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

This driver did not set the MM2S Fs Multiplier Register to the proper
value for playback streams. This needs to be set to the sample rate to
MCLK multiplier, or random stream underflows can occur on the downstream
I2S transmitter.

Store the sysclk value provided via the set_sysclk callback and use that
in conjunction with the sample rate in the hw_params callback to calculate
the proper value to set for this register.

Fixes: 6f6c3c36f091 ("ASoC: xlnx: add pcm formatter platform driver")
Signed-off-by: Robert Hancock <robert.hancock@calian.com>
Link: https://lore.kernel.org/r/20220120195832.1742271-2-robert.hancock@calian.com
Signed-off-by: Mark Brown <broonie@kernel.org>
State: upstream (1c5091fbe7e0d0804158200b7feac5123f7b4fbd)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>