spi: spi-zynqmp-gqspi: Update driver to support both IO and dma modes
commit e14ccf6ec5febb9125b1f4e48748a0645d504920 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Existing driver will support dma and if data bytes are less than 8
then will move to IO mode.
but this patch switches the driver completely to either IO or dma
based on has-io-mode dts property. its a boolean property and if
exists the driver will switch to IO mode.

This is needed for spi-nor ubifs. since spi-nor ubifs won't work
with dma hence we need to switch our driver to operate in IO mode.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>