usb: dwc3: hibernation: fix scratch buffer allocation
commit 18808ad7e10b83acbf4f12cb2694821c67aa85a5 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

DWC3 core fails to save in scratchpad buffer area if it's not
initialized with zero.

We would know how many scratch buffers to allocate in core_init. So,
move scratch buffer allocation in core_init. Keep a check to avoid
re-allocation when core_init is called after probe.

Signed-off-by: Manish Narani <manish.narani@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>