usb: dwc3: xilinx: Add gpio-reset support
commit e00e26200b65ff41d23721bda13ef7115972d039 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

This patch adds a USB GPIO based reset for dwc3-xilinx driver. The PHY
needs to be resetted after completion of phy initialization. As part
of the reset, checks for gpio-reset binding before toggling the pin.
Delay of millisecond is added in between low and high to meet the setup
and hold time requirement of the reset.

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Signed-off-by: Manish Narani <manish.narani@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>