usb: chipidea: udc: make controller hardware endpoint primed
commit f316898af2c9a612695010796abf5a6f04ff6726 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Root-cause:
There is an issue like endpoint is not recognized as primed, when bus
have more pressure and the add dTD tripwire semaphore (ATDTW bit in
USBCMD register) that can cause the controller to ignore a dTD that is
added to a primed endpoint.
This issue observed with the Windows10 host machine.

Workaround:
The software must implement a periodic cycle, and check for each dTD,
if the endpoint is primed. It can do this by reading the corresponding
bits in the ENDPTPRIME and ENDPTSTAT registers. If these bits are read
at 0, the software needs to re-prime the endpoint by writing 1 to the
corresponding bit in the ENDPTPRIME register.

Added conditional revision check of 2.20[CI_REVISION_22].

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Signed-off-by: Manish Narani <manish.narani@xilinx.com>
State: upstream (05735f0854e1e2fc0dd266cc6f583fc79dfdd5d8)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>