LF-5409 video: fbdev: mxc_ipuv3_fb: Explicitly configure pre.field_inverse in _setup_disp_channel2()
commit 3cc1e5a69ce6d67332990fbd0b3eeaa2838d6dee from
https://source.codeaurora.org/external/imx/linux-imx

The uninitialized pre.field_inverse field of struct ipu_pre_context in
_setup_disp_channel2() could contain random data because the compiler
doesn't guarantee to initialize it to zero.  The random data could be
configured into the PRE engine, which may cause unknown behavior of
the hardware.  This patch explicitly configures pre.field_inverse in
every case to avoid the issue.

Cc: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>