arm64: dts: ti: k3-j721s2-main: add wave5 video encoder/decoder node
commit 14c566bd681ad881579e186ec964e2cbff8423f2 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add the Chips and Media wave521cl video decoder/encoder node on J721S2.

This functional block also requires an SRAM buffer as a bandwidth saving
temporary store so we need to add a carve out of 126K for this as
specified in the documentation.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>