net: macb: Added ZynqMP-specific initialization
commit 19309addec7c147eb4419d52b7792f84edf6e9ce from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

The GEM controllers on ZynqMP were missing some initialization steps which
are required in some cases when using SGMII mode, which uses the PS-GTR
transceivers managed by the phy-zynqmp driver.

The GEM core appears to need a hardware-level reset in order to work
properly in SGMII mode in cases where the GT reference clock was not
present at initial power-on. This can be done using a reset mapped to
the zynqmp-reset driver in the device tree.

Also, when in SGMII mode, the GEM driver needs to ensure the PHY is
initialized and powered on.

On top of this mainline patch, resolve merge conflicts and drop
passing of phy-names in obtaining SGMII PHY handle.

Signed-off-by: Robert Hancock <robert.hancock@calian.com>
Reviewed-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Link: https://lore.kernel.org/all/20220127163736.3677478-3-robert.hancock@calian.com
State: waiting (accepted in net-next)
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>