iommu/omap: Add transition support between hwmod and ti,sysc
commit 9f6939aacbe9a0075449eba51dc291226b089ad1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The OMAP IOMMU driver relies on hwmod layers and uses platform data
callbacks for clocking and reset functionality at present. The same
will be achieved through the ti-sysc parent interconnect node in the
future using the pm_runtime framework and the inherent parent-child
relationship.

Add support to the code so that both the newer ti-sysc based logic
and the older hwmod based logic are supported as the IOMMU DT nodes
are transitioned to the ti-sysc mode. The existing platform data
callbacks should continue to be invoked when using hwmods, but
should not be invoked with ti-sysc.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>