net/rpmsg: return ESHUTDOWN upon Tx on errored sockets
commit 1c8f70b02d83c6d1a71e9079805cedffccd04dcd from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The rpmsg proto driver uses a single rpmsg channel device
published from a remote processor to transmit all socket-based
messages intended for that remote processor. This channel will
be auto-removed and recreated if the remote processor goes
through an error recovery process. Any connected sockets are
marked with an error status, and further transmissions on these
connected sockets should gracefully return an error. This error
condition is specifically checked for and a new error ESHUTDOWN
is returned back to userspace to differentiate it from
transmissions on an unconnected socket.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Drew Fustini <dfustini@baylibre.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>