media: i2c: Add support for 19.2MHz clock to ov7251
commit  4d215d67671be7a6f132af7074b2040a3db32da0 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

The OV7251 sensor is used as the IR camera sensor on the Microsoft
Surface line of tablets; this provides a 19.2MHz external clock. Add
the ability to support that rate to the driver by defining a new set
of PLL configs. Extend the clock handling in .probe() to check for
either supported frequency.

Signed-off-by: Daniel Scally <djrscally@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>