firmware: xilinx: fix backward compatibility
commit 3ebad4f5e26a8c38f476112f26b7f0241581857c from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

After enabling the feature check support the backward compatibility
got broken. For earlier releases the feature check support wasn't
there so it shouldn't enable feature check for them.

Fix the same by checking whether the feature check is supported or not
in the firmware.

Also, moving get_set_conduit_method() before as the Linux is
requesting to ATF for the PM_FEATURE_CHECK API version for which the
interface should be enabled with ATF.

Fixes: baf44f71ecf9 ("firmware: xilinx: enable feature check for ZU+")
Signed-off-by: Ronak Jain <ronak.jain@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>