LF-5606: libata: pmp: workaround for ERR051238 for imx sata
commit  35971bf0e5e710ab1e22f20e90e765ff80394ead from
https://github.com/nxp-imx/linux-imx lf-6.1.y

Information about ERR051238:

Title: SATA PMP Hot Plug Async Notification Not Working

Description: A hot plug on PMP port is communicated to host through
Asynchronous notification using “Set Device Bits” (SDB) FIS which
indicate that change has happened on a port of PMP by setting the
SNotify bit along with “interrupt” and “Notify” bits. The requirement
for a host to support Asynchronous Notification is that it may
generate an interrupt if the Set Device Bits FIS is received.
The SNotification register enables software to be sure that the
interrupt was due to a notification event. In this case the Host
controller doesn't set the SNotify bit due to which the Software
is not aware of the Hotplug event.

This patch creates a polling thread to detect the hotplug event.
Once hotplug event was detected, it sets up the flags and notify
the system to take actions manually.

NOTE:
1. Limit it to spawn only one hotplug thread.
2. Don't let the thread call async_notification twice during
   the same hotplug. (add a one-second cooldown)

Signed-off-by: Andy Tang <andy.tang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>