octeontx2-af: modify TL1 rr_prio value from 1 to 9
commit 4845d98be99a3d3aeaaef35ba771d5c34d47eb44 from
git@git.assembla.com:cavium/WindRiver.linux.git

The TL1 queue of a link is currently configured to use rr_prio
as 1. This would conflict with the use case in which kernel
traffic must be prioritised over other traffic. Since the
hardware scheduler traverses the TL2 child nodes beginning
with TL1 prio_anchor + {0...9}, it is necessary to provide
low priority DWRR child nodes final priority value so that
kernel PF/VFs may be placed at the beginning of the list.

Change-Id: Iecbce283371a194a094e037d4d24d653f8f2a1dd
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/72134
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>