arm64: zynqmp: Comment sata phy setting for zcu102
commit 6d1044b1413f30b6a08279ee4fdaed9542bcab99 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Debian, Fedora are not able to handle GTR run time setup because drivers
for PHY driver and si5341 driver are not enabled by default. That's why for
Systemready IR certification there is a need to comment this run time
setting. GTR lines are setup already by FSBL that's why it is not big
problem.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>