LF-5643-1 ASoC: SOF: imx8m: Fix SOF DSP driver probe
commit 16054e38278358fe058ca38bbd9fb6ef4aed3454 from
https://github.com/nxp-imx/linux-imx.git

Commit 9ba23717b29270 ("ASoC: SOF: imx8m: Implement DSP start") used
fsl,dsp-ctrl for audiomix compatible named instead of "fsl,imx8mp-audio-blk-ctrl".

Actually, the intention of the change was good but was incomplete.

Because of 61f3c905e66e1 ("LF-4209-6: arm64: dts: imx8mp: update DSP")
we already have a reference to audiomix register via fsl,dsp-ctrl so
we no longer need to get the audiomix regmap via compatible string but
get it via phandle.

So, in this patch we will retrieve theaudiomix regmap via phandle. In a
subsequent patch we will get rid of audiomix-dsp-regmap property since
we already have fsl,dsp-ctrl.

Fixes: 9ba23717b29270 ("ASoC: SOF: imx8m: Implement DSP start")
Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>