ASoC: atmel: mchp-pdmc: do not remove non-runtime controls
commit cd8c7b96ca71b84dbdad3b8053cf4267ec83e160 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

Removing controls causes issues if the stream is started from another
control. Do not remove these controls when the stream is
started/finished and make it return -EBUSY instead. The controller is
considered busy when the audio filter and the SINC filter are used.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@microchip.com>
[claudiu.beznea: adapt to v6.1.4]
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>
(cherry picked from linux-6.1-trunk/at91/sound)
[andrei.simion: conflicts solved and adapt to v6.6.9]
Signed-off-by: Andrei Simion <andrei.simion@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>