media: i2c: Rename ad5398 to ad5398_vcm
commit  bbf89a7d1d964585ca55a6167653a8b8c9650193 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

There's already a regulator module called ad5398 that exposes
this device through the regulator API. That is meaningless in
the terms that it uses and how it maps to V4L2, so a new driver
was added. However the module name collision wasn't noted, so
rename it now.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>