v4l: xilinx: axis-subsetconv: Add driver for axis subset converter
commit 62845b3eb9529e0df12f03484cfb4fa0f798174b from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

This converter driver is for matching the format of source pad and
sink pad in the media pipeline. The format of a source pad does
not match the sink pad if it is converted by a non-memory mapped
hardware IP. This subset converter driver is for non-memory mapped
axi stream subset converter which converts the format of the stream.

Refer to Xilinx PG085 document for IP information.

Signed-off-by: Anil Kumar Mamidala <anil.mamidala@xilinx.com>
Reviewed-by: Vishal Sagar <vishal.sagar@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>