mmc: sdhci-of-arasan: Add delay to reach card stable state
commit 5bec44cdeb9d43cd71bbfec68adf372945a7d212 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Once eMMC controller brings out of reset, it is observed Card Insert and
Card Stable bits are not setting immediately, it is taking 1msec time
to get these two bits set in the hardware.
This behavior is not expected for eMMC because eMMC is an embedded slot,
so there should not be any Card detection logic involved in HW,
Card Insert Card Stable bits should just set within no time.
This patch add delay of 1msec as a workaround after eMMC controller
brings out of reset.

Signed-off-by: Sai Krishna Potthuri <lakshmi.sai.krishna.potthuri@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>