drm/vc4: Make sure we don't end up with a core clock too high
commit  e7facdb533e16623674bd81c9d303657de0f07f3 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

Following the clock rate range improvements to the clock framework,
trying to set a disjoint range on a clock will now result in an error.

Thus, we can't set a minimum rate higher than the maximum reported by
the firmware, or clk_set_min_rate() will fail.

Thus we need to clamp the rate we are about to ask for to the maximum
rate possible on that clock.

Reviewed-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>