drm/panel: panel-ilitek9881c: Add prepare_upstream_first flag
commit  77ef4c1a70f08632c827dbef55c8398adcbf915a from
https://github.com/raspberrypi/linux.git rpi-5.15.y

The panel sends MIPI DCS commands during prepare and is expecting
the bus to remain in LP-11 state in-between.

Set the prepare_upstream_first flag so that the upstream DSI host
is prepared / pre_enabled first, and therefore the bus is in a
defined state.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>