irqchip/gic-v3: Detect LPI invalidation MMIO registers
commit a84b2580a629dbffbfb94d5944d0bedef7cafa0d from
git@git.assembla.com:cavium/WindRiver.linux.git

Since GICv4.1, an implementation can offer the same MMIO-based
implementation as DirectLPI, only with an ITS. Given that this
can be hugely beneficial for workloads that are very LPI masking
heavy (although these workloads are admitedly a bit odd).

Interestingly, this is independent of RVPEI, which only *implies*
the functionnality.

So let's detect whether the implementation has GICR_CTLR.IR set,
and propagate this as DirectLPI to the ITS driver.

While we're at it, repaint the GICv3 banner so that we advertise
the various capabilities at boot time to be slightly less invasive.

Change-Id: Ibfd38fa9e61ec5b63965ad6fac24f48c4600f88d
Signed-off-by: Marc Zyngier <maz@kernel.org>
Link: https://lore.kernel.org/r/20220405183857.205960-3-maz@kernel.org
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/102748
Reviewed-by: Linu Cherian <lcherian@marvell.com>
Tested-by: Linu Cherian <lcherian@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>