HSD #14016448221: fix crypto services firmware loading
commit c972a765dc00d2d3e90ef5217478ab124f841d3b from
https://github.com/altera-opensource/linux-socfpga.git socfpga-6.6.22-lts.

The mainline downstream crypto service layer driver is not getting loaded
correctly because there was an issue with the LTS version getting out sync
with the mainline(i.e. socfpga-5.16). The crypto services driver was
designed as a standalone driver that has it own dts binding
"intel,agilex-soc-fcs", but somehow this patch:

40d2e422ced9d ("firmware: stratix10-svc: extend svc to support FCS features")

had code to platform add the crypto services driver. Remove the platform_add
call for the crypto services driver.

Fixes: 40d2e422ced9d ("firmware: stratix10-svc: extend svc to support FCS features")
Signed-off-by: Dinh Nguyen <dinh.nguyen@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>