net: ethernet: ti: icssg_prueth: fix devlink port register sequence
commit 34ae2df5d2cc7266cf8c84584df53d3ba9cb4971 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When udevd is configured to rename interfaces according to persistent
naming rules and if a network interface has phys_port_name in sysfs, its
contents will be appended to the interface name. However, register_netdev
creates device in sysfs and if devlink_port_register is called after that,
there is a timeframe in which udevd may read an empty phys_port_name value.
The consequence is that the interface will lose this suffix and its name
will not be really persistent.

The solution is to register the port before registering a netdev.

Fixes: 3f7a788ef4c3 ("net: ethernet: ti: icssg_prueth: Add support for ICSSG switch firmware on AM654 PG2.0")
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>