arm64: dts: ti: Add AM62x SK camera overlay
commit d74bb41bf98651e50c37441dcf218ea697f775d2 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add an overlay for OV5640 camera sensor for AM62x SK. The camera module
used is Digilent PCam 5C, which provides OV5640 sensor with the
Raspberry Pi connector.

The module does not work at 400 KHz reliably. Drop the I2C bus frequency
to 10 KHz to make it work properly. The only other peripheral on the bus
is the expansion header so the impact should not be too large.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Tested-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>