media: subdev: use streams in v4l2_subdev_link_validate()
commit f1c83a8c6b86aae69b87236e6467bbfc532e7913 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Update v4l2_subdev_link_validate() to use routing and streams for
validation.

Instead of just looking at the format on the pad on both ends of the
link, the routing tables are used to collect all the streams going from
the source to the sink over the link, and the streams' formats on both
ends of the link are verified.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>