dt-bindings: display: ti,am65x-dss: Fix missing reg and max interrupts
commit dba8b6a1060807a22a564dd79b8b7fad28ef6eba from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The DSS IP on the am65x soc supports in addition to common, a common1
register space as well. Further, the max number of interrupts the DSS
supports is 2 and not 1.

Additionally, this patch reflects the yaml changes in the affected dts
files.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Tested-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>