arm64: dts: ti: k3-am625-sk: Add HDMI tx, connector & master clock nodes
commit 0876df784093632044cd8a33c9592a0b526272c2 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add DT nodes for sil9022 HDMI transmitter (tx), HDMI connector and the
HDMI fixed master clock on the am625-sk board.

The sil9022 HDMI tx is connected on the i2c1 bus. The HDMI connector on
the board is of type "a". The clock frequency of the master clock that
supports the HDMI tx is 12.288 MHz.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Tested-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>