arm64: dts: ti: k3-am625-sk: Add DDR carveout memory for M4F remote processor
commit 2e4804e96701f3e83419aa014b6520e2dc039010 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Two carveout reserved memory nodes have been added for the M4F remote
processor device node. The first region will be used as the DMA pool for
the rproc device, and the second region will furnish the static carveout
regions for the firmware memory.

The current carveout addresses and sizes are defined statically for the
remote device. The M4F processor does not have an MMU, and as such
requires the exact memory used by the firmware to be set-aside. The
firmware images do not require RSC_CARVEOUT entries in their resource
tables either to allocate the memory for firmware memory segments.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>