drm/vc4: Consolidate Hardware Revision Check
commit  34c64d366bafcf88b7476d5d60e0d1f794a348f2 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

A new generation of controller has been introduced with the
BCM2711/RaspberryPi4. This generation needs a bunch of quirks, and over
time we've piled on a number of checks in most parts of the drivers.

All these checks are performed several times, and are not always
consistent. Let's create a single, global, variable to hold it and use
it everywhere.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>