drm/vc4: Warn if some v3d code is run on BCM2711
commit  83f1a9c43d242e6131027197195795187a526f6b from
https://github.com/raspberrypi/linux.git rpi-5.15.y

The BCM2711 has a separate driver for the v3d, and thus we can't call
into any of the driver entrypoints that rely on the v3d being there.

Let's add a bunch of checks and complain loudly if that ever happen.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>