drm/vc4: drv: Skip BO Backend Initialization on BCM2711
commit  84ed11d606218812895b7e541dc59a278753bd05 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

On the BCM2711, we currently call the vc4_bo_cache_init() and
vc4_gem_init() functions. These functions initialize the BO and GEM
backends.

However, this code was initially created to accomodate the requirements
of the GPU on the older SoCs, while the BCM2711 has a separate driver
for it. So let's just skip these calls when we're on a newer hardware.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>