mailbox: ti-msgmgr: Operate mailbox in polled mode during system suspend
commit 7bcd04970855647827a1d04287b1c62fed42e98d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit df227dc8a68d81b7fe3416eca16d1f21d0b537f0 upstream.

During the system suspend path we must set all queues to operate in
polled mode as it is possible for any protocol built using this mailbox,
such as TISCI, to require communication during the no irq phase of suspend,
and we cannot rely on interrupts there.

Polled mode is implemented by allowing the mailbox user to define an
RX channel as part of the message that is sent which is what gets polled
for a response. If polled mode is enabled, this will immediately be
polled for a response at the end of the mailbox send_data op before
returning success for the data send or timing out if no response is
received.

Finally, to ensure polled mode is always enabled during system suspend,
iterate through all queues to set RX queues to polled mode during system
suspend and disable polled mode for all in the resume handler.

Signed-off-by: Jassi Brar <jaswinder.singh@linaro.org>
[dgerlach@ti.com: Switch DEFINE_SIMPLE_DEV_PM_OPS to SIMPLE_DEV_PM_OPS]
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>