firmware: ti_sci: Switch transport to polled mode during system suspend
commit a5be91d495037b884bf7e7bc19ca90c0f8ca42de from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

During system suspend it is completely valid for devices to invoke TISCI
commands during the noirq phase of the suspend path. Specifically this
will always be seen for devices that define a power-domains DT property
and make use of the ti_sci_pm_domains genpd implementation.
The genpd_finish_suspend call will power off devices during the noirq
phase, which will invoke TISCI.

In order to support this, the ti_sci driver must switch to not use
wait_for_completion_timeout during suspend, but instead rely on a manual
check for if the completion is not yet done, and proceed only if this is
the case.

[dgerlach@ti.com: Switch DEFINE_SIMPLE_DEV_PM_OPS to SIMPLE_DEV_PM_OPS]
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>