firmware: ti_sci: Introduce ti,ctx-memory-region for reserved LPM memory
commit e5f01f73e0b4203ff6611f6d2b46513b50cc36a0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

A reserved memory region in DDR must be used during low power modes for
saving of some system context when using the ti_sci firmware, so
introduce a property to allow providing this in the device tree so that
it can be read and shared with the firmware.

Also send a TISCI_MSG_PREPARE_SUSPEND message to the firmware during
probe to determine if system suspend is supported and if
ti_sci_init_suspend should be called based on the response received.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>