firmware: ti_sci: Use dt provided fw name and address to load at suspend time
commit 8a2a2afc3bc91bb170d4cd673826ed0459578c59 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Use request_firmware_direct to load the fs stub LPM firmware to a
provided lpm memory region. The filename for the firmware is provided in
the device tree as "ti,lpm-firmware-name"

Also, add support for "lazy loading" of this firmware during the suspend
callback for the driver so that it is loaded at first suspend. It is
possible in the future for this firmware to require reload, so add a
check to indicate that the firmware is currently loaded so it is only
loaded once at this time.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>