LF-5945: LF-5924: imx8m: vpu_dec: report source change event when bit depth is changed
commit e90a04ddda15d8497e37fe2c4c1b829f0273b2fc from
https://github.com/nxp-imx/linux-imx.git

Driver will report all possible format in enum_fmt
before source change event,
so client may choose the 10bit format, for example nv12_10le.
but if the stream is 8bit, when the sequence header is parsed,
only 8 bit format is supported,
so need to send source change event to client,
then client can setup capture with 8bit format.

If the source change event is missed, the result may be unexpected

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>