LF-5945-2: imx8m: vpu_dec: remove check reschange in driver
commit d0502afc875e23ce5718e1f88187734249b61919 from
https://github.com/nxp-imx/linux-imx.git

after daemon send reschange event to driver,
it requires driver to send captureon cmd to it,
otherwise it keeps waiting.
we add this check to compatible with ffmpeg.
if the source change event is handled by capture streamon
or start cmd,
driver can always notify userspace a source change event.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>