drivers/perf: Add PMU to support reading resctrl counters via perf
commit e551bd411852030daf3536cc5a210cd1f24a3a2d from
git@git.assembla.com:cavium/WindRiver.linux.git

Resctrl exposes the cache opccupancy and bandwidth counters to user space
via files. Each control and monitor group has a copy of these files, and
can read the counter by reading the file.

MPAM needs to allocate a hardware monitor to do the counting work, and
there may not be enough for every control and monitor group to have one.
MPAM needs some indication of which counters are currently in use, it
needs to be able to reject additional users if there are not enough
monitors.

Add a PMU driver for resctrl. This never touches the hardware directly,
it only uses the resctrl API to retrieve the counter values. The PMU
supports the same events as resctrl. The event parameters are the domain
id, which user-space already understands, and the control or monitor
group id. The group id is a 64bit field provided by resctrl that can
be used in the kernel to retrieve the group. Currently this encodes
the CLOSID and RMID.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>