x86/resctrl: Move mbm_state updates to __rmid_read()
commit c386aeaceb06cf150ecd81aedcc887da4afb2661 from
git@git.assembla.com:cavium/WindRiver.linux.git

__rmid_read() may be called via IPI, because of this it does the minimum
amount of work. The updates to the mbm_state are done by the caller.

To support the PMU driver calling resctrl_arch_rmid_read() in irq
context, this code needs to be able to restart if was called in an
interrupt.

Move the mbm_state updates to __rmid_read(), this makes it easier
to restart, without sending a second IPI.

No change in behaviour.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>